Safari-609.1.11.1/LayoutTests/js/arrowfunction-associativity-2-expected.txt

Tests for ES6 arrow function nested declaration

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


af = a => b => a
PASS af('ABC')('DEF') is 'ABC'
PASS successfullyParsed is true

TEST COMPLETE
